[ADD] add balance

+++ controllers/expense.js
@@ -3,12 +3,15 @@ const ExpenseSchema = require("../models/expenseModel")
 exports.addExpense = async (req, res) => {
   const {title, amount, category, description, date}  = req.body
 
+  const userId = req.user.userId;
+
   const expense = ExpenseSchema({
     title,
     amount,
     category,
     description,
-    date
+    date,
+    userId
   })
 
   try {
@@ -21,30 +24,34 @@ exports.addExpense = async (req, res) => {
     await expense.save()
     res.status(200).json({ message: "expense added successfully" })
   } catch (error) {
-    res.status(500).json({ message: "Server error" })
-    
+    res.status(500).json({ message: "Server error" })    
   }
-
-  console.log(expense)
 }
 
 
 exports.getExpenses = async (req, res) => {
   try {
-    const expenses = await ExpenseSchema.find().sort({createdAt: -1 })
+    const expenses = await ExpenseSchema.find({ userId: req.user.userId }).sort({createdAt: -1 })
     res.status(200).json(expenses)
   } catch (error) {
     res.status(500).json({ message: "Server error" })
   }
 }
 exports.deleteExpense = async (req, res) =>{
-    const {id} = req.params;
-    ExpenseSchema.findByIdAndDelete(id)
-        .then((expense) =>{
-            res.status(200).json({message: 'expense Deleted'})
-        })
-        .catch((err) =>{
-            res.status(500).json({message: 'Server Error'})
-        })
+    try {
+        const { id } = req.params;
+
+        const expense = await ExpenseSchema.findOne({ _id: id, userId: req.user.userId });
+
+        if (!expense) {
+            return res.status(404).json({ message: 'Income not found or unauthorized' });
+        }
+
+        await ExpenseSchema.findByIdAndDelete(id);
+
+        res.status(200).json({ message: 'Income Deleted' });
+    } catch (err) {
+        res.status(500).json({ message: 'Server Error' });
+    }
 }
